move db into folder
