+++ server/api/users.js
@@ -1,9 +1,9 @@
-var express = require('express')
-var router = express.Router()
+const express = require('express')
+const router = express.Router()
 
 /* GET users listing. */
 router.get('/', (req, res, next) => {
-  res.send('respond with a resource')
+  res.json({ msg: 'users' })
 })
 
 module.exports = router
